feat: Finished dependent child credit
