+++ changelog_entry.yaml
@@ -1,1 +1,4 @@
-
+- bump: minor
+  changes:
+    added:
+    - Dependent child credit program
